Notification Update 4 and Crash Report Added
1. Notification priority set to HIGH for heads-up notifications. The FCM
notification messaging service is called in the AccountActivity classes
to start background service as soon as app is opened.
2. Code for service restart added to service class to restart actvity if
the task is removed and Android OS tries to kill the service. (Some
devices will not recieve notifications if the app is removed from recent
apps as their OS force stops and kills these apps and the services too).
3. Firebase Crash Report dependency added to gradle to recieve crash
reports on the firebase dashboard if the app crashes henceforth.
4. Crash fix for UserRejectSinglePost and code for onBackPressed added.